+++ app/main.py
@@ -24,7 +24,7 @@ def some_stuff():
 
 def chessboard_stuff():
 	app = WallEApp()
-	app.save_chessboards()
+	app.save_solidcolors()
 	pass